+++ index.d.ts
@@ -1,5 +1,8 @@
+import { Transform as TransformClass } from "stream";
+import { Obfuscator as ObfuscatorClass } from "./src/index";
 import {
   IJsConfuser as JsConfuser,
+  IJsConfuserDebugTransformations,
   IJsConfuserObfuscate,
   IJsConfuserPresets,
 } from "./src/types";
@@ -7,3 +10,6 @@ import {
 export default JsConfuser;
 export const obfuscate: IJsConfuserObfuscate;
 export const presets: IJsConfuserPresets;
+export const debugTransformations: IJsConfuserDebugTransformations;
+export const Obfuscator: typeof TransformClass;
+export const Transform: typeof ObfuscatorClass;